% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpt_meta.r
\name{adpt_meta}
\alias{adpt_meta}
\title{Adapt meta information inside a nlmixr UIF}
\usage{
adpt_meta(mdl, newvals)
}
\arguments{
\item{mdl}{character with the name of the model to adapt}

\item{newvals}{list with characteristics/meta data to adapt}
}
\value{
character vector with model including the adapted meta data
}
\description{
regular expressions are used to search for meta data inside a model
file. This meta data is then updated with the provided new values
}
\examples{

\dontrun{
  adpt_meta("model.r",newvals=list(imp=4,ref="run 1"))
}
}
\author{
Richard Hooijmaijers
}
