% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proj.r
\name{create_proj}
\alias{create_proj}
\title{Creates a new project}
\usage{
create_proj(loc = ".", overwrite = FALSE)
}
\arguments{
\item{loc}{character with the location where the project should be created}

\item{overwrite}{logical indicating if files should be overwritten if already exists}
}
\value{
nothing will be returned by the function (only system commands are issued)
}
\description{
Creates a new project which basically means that within the specified folder,
the necessary folder structure will be created and some example models will be placed in it.
}
\examples{

\dontrun{
 create_proj()
}
}
\author{
Richard Hooijmaijers
}
