% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreplot.r
\name{exploreplot}
\alias{exploreplot}
\title{Function to create plot from data exploration app}
\usage{
exploreplot(inputlist)
}
\arguments{
\item{inputlist}{list with input items to create a plot}
}
\value{
a character string with the ggplot code
}
\description{
This function creates a text string for a ggplot based on an input list. This function
is specifically written to be used with the shiny app for data exploration.
}
\examples{

\dontrun{exploreplot(input)}
}
\author{
Richard Hooijmaijers
}
\concept{Plotting functions}
