% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.r
\name{get_meta}
\alias{get_meta}
\title{Get the meta data out of a model function}
\usage{
get_meta(mdl)
}
\arguments{
\item{mdl}{character with the path of the model function}
}
\value{
A list with the models meta data
}
\description{
This function gets only the meta data from a function
}
\examples{

\dontrun{
  get_meta("run1.r")
}
}
