% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proj.r
\name{get_proj}
\alias{get_proj}
\title{Read in and update model results in project object}
\usage{
get_proj(moddir = "./models", proj = "./shinyMixR/project.rds",
  datdir = "./data", geteval = TRUE)
}
\arguments{
\item{moddir}{character with the directory that includes the model files}

\item{proj}{character with the rds files that includes the model information}
}
\description{
This function creates or updates a project object with models and/or results emerged from nlmixr runs
A check is performed to see if newer results are present and only updates these.
}
\examples{

\dontrun{
  proj <- get_proj()
}
}
