% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_mdl.r
\name{incr_mdl}
\alias{incr_mdl}
\title{Increments a model name}
\usage{
incr_mdl(mod, checkloc = NULL)
}
\arguments{
\item{mod}{character with the model name}

\item{checkloc}{character with the location to check for existence of a file}
}
\value{
character with the incremented name
}
\description{
A model name is incremented either by incrementing numerical or alpha numerical.
Furthermore it is possible to check the existence of the incremented model
and take this into account.
}
\examples{

 incr_mdl("run01.r")
}
\author{
Richard Hooijmaijers
}
