% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataexplore.R
\name{module_dataexplore_server}
\alias{module_dataexplore_server}
\title{Data exploration module for server}
\usage{
module_dataexplore_server(id, tabswitch)
}
\arguments{
\item{id}{Module id}

\item{tabswitch}{Reactive value that monitors the tabswitch}
}
\description{
Data exploration module for server
}
