% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataexplore.R
\name{module_dataexplore_ui}
\alias{module_dataexplore_ui}
\title{Data exploration module for UI}
\usage{
module_dataexplore_ui(id)
}
\arguments{
\item{id}{Module id}
}
\description{
Shiny module for data exploration
}
