% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_edit.R
\name{module_edit_server}
\alias{module_edit_server}
\title{Editor module for server}
\usage{
module_edit_server(id, r, settings)
}
\arguments{
\item{id}{Module id}

\item{r}{reactive values object that is defined top-level}

\item{settings}{reactive value with the app settings}
}
\description{
Editor module for server
}
