% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_edit.R
\name{module_edit_ui}
\alias{module_edit_ui}
\title{Editor module for UI}
\usage{
module_edit_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for model editor
}
