% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_fitplots.R
\name{module_fitplots_ui}
\alias{module_fitplots_ui}
\title{Fit plots module for UI}
\usage{
module_fitplots_ui(id, proj_obj)
}
\arguments{
\item{id}{Module id}

\item{proj_obj}{Project object}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for fit plots
}
