% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gof.R
\name{module_gof_server}
\alias{module_gof_server}
\title{GOF plots module for server}
\usage{
module_gof_server(id, tabswitch, settings)
}
\arguments{
\item{id}{Module id}

\item{tabswitch}{reactive value that monitors the tabswitch}

\item{settings}{reactive value with the app settings}
}
\description{
GOF plots module for server
}
