% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_metadata.R
\name{module_metadata_server}
\alias{module_metadata_server}
\title{meta data module for server}
\usage{
module_metadata_server(
  id,
  type,
  selline = NULL,
  sellmod = NULL,
  sellcont = NULL,
  r
)
}
\arguments{
\item{id}{Module id}

\item{type}{character with the type of action (either "save" or "overview")}

\item{selline}{reactive with the selected line for a model (for type "overview")}

\item{sellmod}{reactive with the selected model (for type "save")}

\item{sellcont}{reactive with the content of the selected model (for type "save")}

\item{r}{reactive values object that is defined top-level}
}
\description{
meta data module for server
}
