% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_metadata.R
\name{module_metadata_ui}
\alias{module_metadata_ui}
\title{metadata module for UI}
\usage{
module_metadata_ui(id, type)
}
\arguments{
\item{id}{Module id}

\item{type}{character with the type of button to present (either "save" or "overview")}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for meta data
}
