% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview.R
\name{module_overview_server}
\alias{module_overview_server}
\title{Overview module for server}
\usage{
module_overview_server(id)
}
\arguments{
\item{id}{Module id}
}
\description{
Overview module for server
}
