% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview.R
\name{module_overview_ui}
\alias{module_overview_ui}
\title{Overview module for UI}
\usage{
module_overview_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for overview
}
