% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_partable.R
\name{module_pt_server}
\alias{module_pt_server}
\title{Parameter table module for server}
\usage{
module_pt_server(id, r)
}
\arguments{
\item{id}{Module id}

\item{r}{reactive values object that is defined top-level}
}
\description{
Parameter table module for server
}
