% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_partable.R
\name{module_pt_ui}
\alias{module_pt_ui}
\title{Parameter table module for UI}
\usage{
module_pt_ui(id, proj_obj)
}
\arguments{
\item{id}{Module id}

\item{proj_obj}{Project object}
}
\description{
Shiny module for parameter table
}
