% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reports.R
\name{module_reports_server}
\alias{module_reports_server}
\title{Reporting module for server}
\usage{
module_reports_server(id)
}
\arguments{
\item{id}{Module id}
}
\description{
Reporting module for server
}
