% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reports.R
\name{module_reports_ui}
\alias{module_reports_ui}
\title{Reporting module for UI}
\usage{
module_reports_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for reporting
}
