% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_run.R
\name{module_run_server}
\alias{module_run_server}
\title{Run model module for server}
\usage{
module_run_server(id, tabswitch, r)
}
\arguments{
\item{id}{Module id}

\item{tabswitch}{reactive value that monitors the tabswitch}

\item{r}{reactive values object that is defined top-level}
}
\description{
Run model module for server
}
