% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_run.R
\name{module_run_ui}
\alias{module_run_ui}
\title{Run model module for UI}
\usage{
module_run_ui(id, proj_obj)
}
\arguments{
\item{id}{Module id}

\item{proj_obj}{Project object}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for running models
}
