% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scripts.R
\name{module_scripts_server}
\alias{module_scripts_server}
\title{Run script module for server}
\usage{
module_scripts_server(id, files = NULL, loc = "temp", r)
}
\arguments{
\item{id}{Module id}

\item{files}{character vector of files to apply the scripts on, usually a reactive}

\item{loc}{character with the location where the temp scripts are saved (created when not existing)}

\item{r}{reactive values object that is defined top-level}
}
\description{
Run script module for server
}
