% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scripts.R
\name{module_scripts_ui}
\alias{module_scripts_ui}
\title{Run script module for UI}
\usage{
module_scripts_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for running scripts
}
