% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{module_settings_server}
\alias{module_settings_server}
\title{Settings module for server}
\usage{
module_settings_server(id)
}
\arguments{
\item{id}{Module id}
}
\description{
Settings module for server
}
