% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{module_settings_ui}
\alias{module_settings_ui}
\title{Settings module for UI}
\usage{
module_settings_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for settings
}
