% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{myalert}
\alias{myalert}
\title{wrapper function for sweetalert in shinywidgets}
\usage{
myalert(text, type, ...)
}
\arguments{
\item{text}{character with the text to display}

\item{type}{character with the type of alert to display}

\item{...}{other arguments passed to class}
}
\value{
No return value, called for side effects
}
\description{
This function gets list of widgets to include in run_shinymixr
}
