% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_shinymixr_test}
\alias{setup_shinymixr_test}
\title{Record a shinyMixR test}
\usage{
setup_shinymixr_test(dir = "./tests/files", overwrite = TRUE, record = FALSE)
}
\arguments{
\item{dir}{character, the directory where the files will be placed and the test will be recorded}

\item{overwrite}{logical, if TRUE, the function will remove the content of the tests/files directory}
}
\value{
runs the shinyMixR interface and records actions in the app
}
\description{
Record a shinyMixR test
}
\keyword{internal}
