% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMixR-package.R
\docType{package}
\name{shinyMixR-package}
\alias{shinyMixR}
\alias{shinyMixR-package}
\title{shinyMixR: Interactive 'shiny' Dashboard for 'nlmixr2'}
\description{
An interface for the 'nlmixr2' package. Furthermore additional functions are included to work with the 'nlmixr2' package through the command line.
}
\author{
\strong{Maintainer}: Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Teun Post \email{teunpost@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item LAPP Consultants \email{info@lapp.nl} [funder, copyright holder]
  \item Matthew Fidler [contributor]
  \item Veerle van Leemput [contributor]
}

}
\keyword{internal}
