% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymixr_gadget.R
\name{shinymixr_gadget}
\alias{shinymixr_gadget}
\title{Rstudio gadget to select project and start app}
\usage{
shinymixr_gadget()
}
\value{
No return value, runs a gadget to start the shinyMixR interface
}
\description{
Rstudio gadget to select project and start app
}
\examples{

\dontrun{
 if (interactive()) shinymixr_gadget()
}
}
\author{
Richard Hooijmaijers
}
