% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{sigdigs}
\alias{sigdigs}
\title{set significant digits without rounding higher numbers}
\usage{
sigdigs(x, sdig = 3)
}
\arguments{
\item{x}{a numerical vector}

\item{sdig}{a single number defining the number of significant digits}
}
\value{
A character vector with formatted numbers
}
\description{
This function sets significant digits without rounding any numbers
}
