% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{theme_shinyMixR}
\alias{theme_shinyMixR}
\title{theme for ggplot output in the shinyMixR package}
\usage{
theme_shinyMixR(fontsize = 12)
}
\arguments{
\item{fontsize}{numeric with the default fontsize passed through to theme}
}
\value{
A list with ggplot theme elements
}
\description{
This function provides a custom theme for ggplot output
}
