% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_overview.r
\name{tree_overview}
\alias{tree_overview}
\title{Creates tree overview of models}
\usage{
tree_overview(proj_obj, ...)
}
\arguments{
\item{proj_obj}{a project object created with \code{\link{get_proj}}}

\item{...}{additional arguments passed to \code{\link{overview}}}
}
\value{
a data frame is returned with the overview
}
\description{
Create a graphical collapsible tree overview of the models within a project.
This is mostly relevant in case the reference of models is included to
visualise the relationship between models
}
\examples{

\dontrun{
 tree_overview(proj_obj)
}
}
\seealso{
\code{\link[collapsibleTree]{collapsibleTreeNetwork}} which does most of the work
}
\author{
Richard Hooijmaijers
}
