% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inits.r
\name{update_inits}
\alias{update_inits}
\title{Update initial estimates from a final model run}
\usage{
update_inits(mod, res, out)
}
\arguments{
\item{mod}{character with the entire model function included}

\item{res}{character with the path to the model result RDS which holds the final estimated}

\item{out}{character with the path for the updated model to save}
}
\value{
nothing will be returned the function saves the updated model to disk
}
\description{
This function update the initial estimates from a model using the final estimates
from a model result file. Currently this function assumes all models were submitted
using shinyMixR opposed to vanilla nlmixr
}
\examples{

\dontrun{
  update_inits(readLines("run2.r"),"shinyMixR/run2.res.rds","run3.r")
}
}
\author{
Richard Hooijmaijers
}
