% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proj.r
\name{get_proj}
\alias{get_proj}
\title{Read in and update model results in project object}
\usage{
get_proj(projloc = ".", geteval = TRUE)
}
\arguments{
\item{projloc}{character with the base location of the shinyMixR project}

\item{geteval}{logical indicating if the model functions should be evaluated}
}
\value{
A named list with information for each model in the 'projloc'
}
\description{
This function creates or updates a project object with models and/or results emerged from nlmixr2 runs.
A check is performed to see if newer results are present and only updates these.
}
\examples{

\dontrun{
  proj <- get_proj()
}
}
