% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataexplore.R
\name{module_dataexplore_server}
\alias{module_dataexplore_server}
\title{Data exploration module for server}
\usage{
module_dataexplore_server(id, r)
}
\arguments{
\item{id}{Module id}

\item{r}{reactive values object that is defined top-level}
}
\value{
No return value, called for side effects
}
\description{
Data exploration module for server
}
