% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_fitplots.R
\name{module_fitplots_server}
\alias{module_fitplots_server}
\title{Fit plots module for server}
\usage{
module_fitplots_server(id, tabswitch, settings)
}
\arguments{
\item{id}{Module id}

\item{tabswitch}{reactive value that monitors the tabswitch}

\item{settings}{reactive value with the app settings}
}
\description{
Fit plots module for server
}
