% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gof.R
\name{module_gof_ui}
\alias{module_gof_ui}
\title{GOF plots module for UI}
\usage{
module_gof_ui(id, proj_obj)
}
\arguments{
\item{id}{Module id}

\item{proj_obj}{Project object}
}
\value{
A list of html tags used for th UI of the app
}
\description{
Shiny module for GOF plots
}
