% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numfmt}
\alias{numfmt}
\title{set significant digits without rounding higher numbers}
\usage{
numfmt(x, sdig = 3, snc = 6)
}
\arguments{
\item{x}{a numerical vector}

\item{sdig}{a single number defining the number of significant digits}

\item{snc}{a single number defining the scientific notation cutoff 
(higher means notation is only used for very small or very large numbers)}
}
\value{
a character vector with formatted numbers
}
\description{
This function sets significant digits without rounding any numbers
}
\examples{
numfmt(c(0.012,12345,1))
}
\author{
Richard Hooijmaijers
}
