% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.r
\name{overview}
\alias{overview}
\title{Creates model overview}
\usage{
overview(proj_obj, ...)
}
\arguments{
\item{proj_obj}{a project object created with \code{\link{get_proj}}}

\item{...}{additional arguments passed to \code{\link{get_proj}}}
}
\value{
a data frame is returned with the overview
}
\description{
Create an overview of the models within a project. This overview includes the meta data
of the models and if results are available, also the objective function and run-times
}
\examples{

\dontrun{
 overview(proj_obj)
}
}
\author{
Richard Hooijmaijers
}
