% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinymixr.r
\name{run_shinymixr}
\alias{run_shinymixr}
\title{Creates and run the interface}
\usage{
run_shinymixr(wd = getwd(), ...)
}
\arguments{
\item{wd}{character with the working directory}

\item{...}{arguments passed to the shiny runApp function}
}
\value{
No return value, runs the shinyMixR interface
}
\description{
Creates and run the interface
}
\examples{

\dontrun{
 if (interactive()) run_shinymixr(".")
}
}
\author{
Richard Hooijmaijers
}
