% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinymixr.r
\name{run_shinymixr}
\alias{run_shinymixr}
\title{Runs the shiny app for nlmixr}
\usage{
run_shinymixr(wd = getwd(), ...)
}
\arguments{
\item{wd}{character with the working directory to start from (wd will be set to this location)}

\item{...}{arguments passed to the shiny runApp function}
}
\description{
Runs the shiny app for nlmixr
}
\examples{

\dontrun{
 run_shinymixr()
}
}
\author{
Richard Hooijmaijers
}
