% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shinymixr.r
\name{run_shinymixr}
\alias{run_shinymixr}
\title{Creates and run the interface}
\usage{
run_shinymixr(wd = getwd(), dry_run = FALSE, ...)
}
\arguments{
\item{wd}{character with the working directory}

\item{dry_run}{logical, if TRUE, the function will not launch the app, but will only create the necessary files}

\item{...}{arguments passed to the shiny runApp function}
}
\value{
runs the shinyMixR interface
}
\description{
Creates and run the interface
}
\examples{

\dontrun{
 run_shinymixr(".")
}
}
\author{
Richard Hooijmaijers
}
