% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMixR-package.R
\docType{package}
\name{shinyMixR-package}
\alias{shinyMixR}
\alias{shinyMixR-package}
\title{shinyMixR: Interactive 'shiny' Dashboard for 'nlmixr2'}
\description{
An R shiny user interface for the 'nlmixr2' (Fidler et al (2019) \doi{10.1002/psp4.12445}) package, designed to simplify the modeling process for users. Additionally, this package includes supplementary functions to further enhances the usage of 'nlmixr2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RichardHooijmaijers/shinyMixR}
  \item \url{https://richardhooijmaijers.github.io/shinyMixR}
  \item Report bugs at \url{https://github.com/RichardHooijmaijers/shinyMixR/issues}
}

}
\author{
\strong{Maintainer}: Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Teun Post \email{teunpost@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item LAPP Consultants \email{info@lapp.nl} [funder, copyright holder]
  \item Matthew Fidler [contributor]
  \item Veerle van Leemput [contributor]
}

}
\keyword{internal}
